refactor(ari): move `RequestClient` to its own file
