+++ arirs/src/lib.rs
@@ -1,3 +1,6 @@
+mod request_client;
+pub use request_client::RequestClient;
+
 mod bridge;
 pub use bridge::Bridge;
 
@@ -5,7 +8,7 @@ mod channel;
 pub use channel::*;
 
 mod client;
-pub use client::{Client, RequestClient};
+pub use client::Client;
 
 mod device;
 pub use device::{DeviceState, DeviceStateChanged};
